ogv1.0
- 为翻译包详情页大幅度修改了页面逻辑，使openGraph能够有效使用
